Aggiunte animazione nell'addView Aggiunta noTaskView implementata logica nella ListView2

+++ Elena/Views/ListView2.swift
@@ -28,19 +28,28 @@ struct ListView2: View {
      */
     
     var body: some View {
-        List{
-            ForEach(listViewModel.items){ item in
-                ListRowView(item: item)
-                    .onTapGesture {
-                        withAnimation(.linear){
-                            listViewModel.updateItem(item: item)
-                        }
+        ZStack{
+            if listViewModel.items.isEmpty {
+                NoTasksView()
+                    .transition(AnyTransition.opacity
+                        .animation(.easeIn))
+            } else {
+                List{
+                    ForEach(listViewModel.items){ item in
+                        ListRowView(item: item)
+                            .onTapGesture {
+                                withAnimation(.linear){
+                                    listViewModel.updateItem(item: item)
+                                }
+                            }
                     }
+                    .onDelete(perform: listViewModel.deleteItem)
+                    .onMove(perform: listViewModel.moveItem)
+                }
+                .listStyle(PlainListStyle())
             }
-            .onDelete(perform: listViewModel.deleteItem)
-            .onMove(perform: listViewModel.moveItem)
         }
-        .listStyle(PlainListStyle())
+        
         .navigationTitle("Daily Tasks 📝")
         .navigationBarItems(
             leading: EditButton(),
